Added Input Polling and now Win32 has an attribute for WNDCLASSEX, removed some cast warnings

+++ Satoshi/src/Satoshi/ImGui/ImGuiLayer.cpp
@@ -65,7 +65,7 @@ namespace Satoshi {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application* app = Application::GetInstance();
-		io.DisplaySize = ImVec2(app->GetWindowWidth(), app->GetWindowHeight());
+		io.DisplaySize = ImVec2((float) app->GetWindowWidth(), (float) app->GetWindowHeight());
 
 		float time = app->GetWindowTime();
 		io.DeltaTime = m_Time > 0.0f ? (time - m_Time) : (1.0f / 60.0f);
@@ -161,7 +161,7 @@ namespace Satoshi {
 	bool ImGuiLayer::OnWindowResizeEvent(WindowResizeEvent& e)
 	{
 		ImGuiIO& io = ImGui::GetIO();
-		io.DisplaySize = ImVec2(e.GetWidth(), e.GetHeight());
+		io.DisplaySize = ImVec2((float) e.GetWidth(), (float) e.GetHeight());
 		io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 		//glViewport(0, 0, e.GetWidth(), e.GetHeight());
